Remove "Contact your representatives" middle column
As discussed at the meeting on Friday 14th October, there is little
need to encourage people to contact their representatives as the
relevent authorities (council, Screen Scotland, MSPs) are already
keenly aware of the issues.

As such, there is no need to have a section on the website to
encourage people to contact their representatives.